clean up keys
