+++ lib/keys/index.js
@@ -1,10 +1,10 @@
 'use strict'
 const filters = require('./filters')
+const instances = require('./instances')
+const keys = require('./method')
 const addFilters = filters.add
 const removeFilters = filters.remove
-const createFilters = filters.create
 const filterTypes = filters.types
-const empty = []
 
 exports.properties = { keyType: true }
 
@@ -18,6 +18,7 @@ exports.define = {
     writable: true
   },
   filterTypes,
+  keys,
   addKey (key, target) {
     var keys = this._keys
     if (!this._ownKeys) {
@@ -38,7 +39,7 @@ exports.define = {
         addFilters(this, key)
       }
     }
-    eachInstance(this, target, this.addKey, keys, key)
+    instances(this, target, this.addKey, keys, key)
   },
   removeKey (key, target) {
     var keys = this._keys
@@ -83,7 +84,7 @@ exports.define = {
         }
       }
     }
-    eachInstance(this, target, this.removeKey, keys, key)
+    instances(this, target, this.removeKey, keys, key)
   },
   removeAllKeys () {
     // removeAllKeys will not update instances
@@ -92,56 +93,5 @@ exports.define = {
     }
     this._filters = false
     this._keys = false
-  },
-  keys (type) {
-    const keys = this._keys
-    if (!keys) {
-      return empty
-    } else {
-      if (type) {
-        createFilters(this, type)
-        const filters = this._filters
-        if (!(type in filters)) {
-          let filter = this.filterTypes
-          let filtered = filters[type] = []
-          for (let i = 0, len = keys.length; i < len; i++) {
-            if (filter(this, keys[i], type)) {
-              filtered.push(keys[i])
-            }
-          }
-          return filtered
-        } else {
-          return filters[type]
-        }
-      } else if ('filter' in this) {
-        createFilters(this, 'keys')
-        const filters = this._filters
-        if ('keys' in filters) {
-          return filters.keys || empty
-        } else {
-          let filtered = filters.keys = []
-          for (let i = 0, len = keys.length; i < len; i++) {
-            if (this.filter(keys[i])) {
-              filtered.push(keys[i])
-            }
-          }
-          return filtered
-        }
-      } else {
-        return keys
-      }
-    }
-  }
-}
-
-function eachInstance (parent, target, method, keys, key) {
-  const instances = parent._instances
-  if (instances) {
-    for (let i = 0, len = instances.length; i < len; i++) {
-      let instance = instances[i]
-      if (instance._keys !== keys && instance[key] === target) {
-        method.call(instance, key, target)
-      }
-    }
   }
 }
